Refactored the nav card I was using on the training page into it's own component to use it for roll-of-honour
