+++ src/components/card/cardPreviews.js
