+++ src/pages/training/index.js
@@ -2,48 +2,14 @@ import React from "react"
 import Layout from "../../components/layout"
 import SEO from "../../components/seo"
 import styled from "styled-components"
-import { graphql, Link } from "gatsby"
-import { getImage, GatsbyImage } from "gatsby-plugin-image"
-
-const StyledLink = styled(Link)`
-  border-radius: 3px;
-  box-shadow: rgba(0, 0, 0, 0.2) 2px 5px 15px;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  min-width: 90vw;
-  max-width: 500px;
-  padding: 0.5rem 1rem;
-  margin: 0.5rem 0;
-  color: inherit;
-  text-decoration: none;
-  transition: transform 0.2s ease-in-out;
-  overflow: hidden;
-
-  &:hover,
-  &:active,
-  &:focus {
-    transform: scale(1.03);
-  }
-`
-
-const CardHeading = styled.h3`
-  margin: 0.2em 0 0.5em 0;
-`
+import SimpleNavCard from "../../components/card/simpleNavCard"
+import { graphql } from "gatsby"
+import { getImage } from "gatsby-plugin-image"
 
 const CardContainer = styled.div`
   display: flex;
   flex-wrap: wrap;
 `
-
-const NavCard = ({ title, featuredImage, slug }) => {
-  return (
-    <StyledLink to={slug}>
-      <CardHeading>{title}</CardHeading>
-      <GatsbyImage image={featuredImage.image} alt={featuredImage.alt} />
-    </StyledLink>
-  )
-}
 
 const TrainingPage = ({ data }) => {
   const getFeaturedImage = wpPage => {
@@ -67,13 +33,13 @@ const TrainingPage = ({ data }) => {
       <h1>Training</h1>
 
       <CardContainer>
-        <NavCard
+        <SimpleNavCard
           slug="/training/senior"
           featuredImage={seniorPageFeaturedImage}
           title="Senior Training"
         />
 
-        <NavCard
+        <SimpleNavCard
           slug="/training/junior"
           featuredImage={juniorPageFeaturedImage}
           title="Junior Training"
